Directional Sharrow Bug Fix
This is not really a bug, but unwanted behavior. If a directional bike
lane width is >0 the bike lane will appear, but so will the sharrow in
that direction. Now the rule will only show a show a sharrow if there is
a no bike lane in that direction. The default color is still green
though if sharrow is active. In addition, the sharrow stamp is now an
extrusion rather than a translation (looks the same, but shape is
closed).