Bike Box Buffer Adapter & Added Source
One issue that has bugged me for a while, but never really decided on a
solution for was that bike boxes looked good only if you had buffer &
parking protection on. A genuine fix might adapt the gap to move to
match the location of the bike lane, but to maintain simplicity and not
add more rules, the gap is just removed and the stop bar is just
stretched across the entire intersection approach. This at least looks
more intentional than the current alternative.  In addition, the github
source for the rule is added.