Zero Lane Fix for Best Fit Setting- Short Term
So for now, I made it so that if the rare occasion occurs that there are
NO drive lanes (nTotalLanes==0), it defaults to old behavior and exact
lane width is used. In the long term, space should be allocated across
ALL splits (bike lane, parking, transit, buffers) equally with some type
of switch based constant...or something a little more clever. For
now...this will do.