Boolean Removal, Slight Refactor
This is actually a large change because it will break people's street's
on update. All boolean parameters are changed to strings, and Bridge
material is refactored to use deleteUV so that if textures are turned
off, the dimension coloring remains. The boolean parameter update will
break most peoples established streets. Demo Scene is also updated.