Bike LTS Correction
This was a minor bug with the Bicycle Rank (LTS) metric. Changed to
check for widths  greater than 0 feet. It is actually hard to catch
because many configurations would not have this problem. The LTS 2 for
Bike Lanes without parking has bike lanes 5.5 ft or less. By checking
for less than 6 ft, this could make a buffered bike lane on a 35 mph
road look less protected than regular bike lane that is less than 6 ft.
Not a lot of configurations would have this issue.