Bike Reverse Fix and Minor Stop Bar/Switch Edit
These are largely bug fixes. One of them introduced at the last commit,
the other is a very minor bug that impacts connected streets that  user
purposely puts stop bars at two street segments that meet that are not
an intersection. It is used elsewhere to eliminate the last stamps for
bus lanes and bike lanes, so I might alter this down the line. For most
streets it is still consistent.