Added Mixed Traffic Bikes
This modification adds another parameter that controls the % of curb
lane traffic (after buses are allocated) that is taken up by bicylists.
This chain involved connecting the vehicle allocation to the number of
lanes, and enables modification later to bias trucks for example to the
curb lane as well.  In addition, this change ties the number of
bicyclists and pedestrians default numbers to the performance measures
of the street.  Finally, defaults for bike lane color and mixed traffic
stay at zero if the bike lane width on either side is a non-zero value.
Next feature will be shared -bus- bike lanes or a bug fix.