Repaired 2 MTL Files
I noticed this odd bug where two people files did not get their textures
mapped appropriately. I am not sure where or when this needed to be
changed, but apparently 2 of the MTL files had Ni 1.00 parameters
specified on the same line as the file path. I am not sure when or if
the files have changed or if how they are parsed is changed. Regardless,
in CE 2016.1, this fixed the issue.