Revised Bus Lane Stamps
This has bothered me for a while. Originally created to label the street
as a bus lane, I have since felt it would make more sense to conform to
NACTO/MUTCD suggestions for preferential lanes.  Just spent some time in
gimp to increase the text size and change it to BUS ONLY.