Enclose Shape Edits, High LOD modifications
These edits are inspired by a continuing commentary with Bastien Clemot
on Geonet (https://geonet.esri.com/people/BastienClemot). Essentially,
added a new LOD setting that allows dash painted lines (lane dividers)
to be be split shapes that look better at long ranges/angled views. In
addition, sidewalk shapes were enclosed with blank shapes for printing
(X.). This was replaced with enclosed shape which does the same thing,
but makes the shape transparent. I hope that works (sorry Geoff if not).
Past that this commit also refines the centerline setting so that none,
is actually none (Asphalt) and white stripe is used for what none used
to be.