Street_Text.cga Started/Added-Minor Function Adds
Text is something I wanted to add to the rule for a while and a frequent thing I hear demanded from users (how do I label the dimensions of the street etc). So to make the rule more like street mix, I am trying to start supporting text in the rule. I don't know what it will look like, yet but this is what I want to start with. Not sure of the direction entirely yet.