Scale UV Fix for White Line Near Center Split
This was a minor aesthetic "bug". Generally, white lines are not created to delineate queue zones on streets on the lane closest to the centerline/median. This is a minor element, but the fix is basically scaling the UV slightly only on the last lane of the approach. This required adding one more parameter to the approach rule.