Delete Centerline_Color Parameter, Add Replacement Options to Center_Type
This is an attempt to fix previous mistakes during the early days of the CS rule. Generally, the Centerline_Color parameter comes from previous rules that this rule initially modeled, but since then it has become more annoying to keep track of two different parameters that have so much control over the Center Sections look and feel. This will likely be updated further, but this is an "breaking" change in that applying this rule to existing CS rule models will require changing the Center_Type to "Yellow Centerline" / Rule default in the case of streets not using a different Center_Type.