Refactor Asphalt Painted
Similar to the last refactor for Asphalt, this is working on simplifying some of the most important rules in the CS rule. It is a moderate lift, but I have done some brief testing and it seems to work. A new parameter will be likely be added to support labeling.