Temporary 2018 Bug Fix
In 2018, CityEngine had some changes to the CGA compiler checks that flagged a few specific elements of the CS rule. The main changes here are that the default landWidth parameter is set to the shape parameter default of 3.5 and solid color options are removed from the planting thematic function. It is unclear why this clears up the compiler error, contacting ESRI to determine. This is a minor loss in functionality.