Tree Refactor and Handles
This request came from Thomas at Esri. One of the things I got asked
from the UC is whether the trees can have handles. Well, they do now.
This edit is extensive, and adds 4 parameters to the rule. In essence,
there are handle calls to median and sidewalk insertion rules, and you
can increase the height of the default tree size accordingly. Radius is
adjusted based on the Plant_Loader.cga functions. If height is near 0,
the tree is not inserted.