Round Curb Insertion
This is a 2018.1 only fix as it depends on insertAlongUV. Self adjust
rounded curb ends will be added to all streets medians, buffers,
boulevards etc. This edit was more extensive than imagined, and helped
catch a few other bugs in the rule. After some testing, I ironed out
most of the kinks in this edit. That said...it could likely use some
more.