Add Curb Edges to Raised Curbs
This is to create a physical barrier between all raised curbs regardless
of where they end. This might change near the round curbs in the future,
but generally this is preempt comments relating to the rule no longer
being able to 3D Printed or having alternating designs between segments
having odd looking holes in them. Solution is imperfect but simple.