Curbside Management Refactor/Bug Fixes
This is still a work in progress, but I found out quickly there was a
reason why I don't use setback operations anywhere else in the rule.
Curved roads create new shapes (regardless of :/= assignment) that
create odd artifacts on these shapes. This "fixes" the problem, and also
deals with fixing the orientation of the bus stop stamp based on the
side of the street. Still workin on this.