Parking, Curbside Management, and UI Edits
Fixed a small inspector ordering issue with the bike lane attributes,
but more critically transitioned the parking functionality to inserted
objects for high definition parking. As part of a general movement away
from prebaked textures for markings to support more high end  use cases,
parallel parking is now entirely based on inserted objects instead of
parking textures. Angled parking had a bug fixed related to previous
changes, but this made reporting rules for parking more complex. In the
future this will be refactored as angled parking is also transitioned to
inserted objects and/or splits. Generally, this has lead to a
significant improvement in the visual quality of parking lanes and
curbside management strategies along side them.