Minor Carshare Stamp Edit
I want to fix this as a whole later on with a general refactor on what
is considered right vs. left for parking lanes. That will have to wait
till scooters and other features are added. This is easy to fix and
simplify, it just requires changes in quite a few places.