Replace Ranges with Enums, Add Custom Arrows
Ranges were given Enums for categorical variables in recent releases. I
moved all categorical text variables to Enums. As part of this, I
attempted to enable custom arrows at intersections using string lists.
This is a partial success as making the Enum for this "unrestricted",
results in a lost of the drop down entirely. While this could replace
the current functionality, it might cause a lot of confusion. This for
now will be something someone will have to edit the Enum to unrestricted
to, or connect to a database for now.