Semantic
After much consideration, I decided I wanted rainbow roads for machine
vision to be something that is part of the rule. This time the thematic
router is overloaded so one can just use the default of "asphalt" for
everything if they edit the rule and don't care for this setting. It
might be removed in future releases pending feedback.