CityEngine 2019 GLB Vehicle Model Upgrade
These models are provided by the Esri CityEngine team as part of
CityEngine. I have used these models to replace the entire vehicles
folder. Additional models will be replaced by these GLBs with time as
they simplify the file structure, is cleaner, and easier to edit the
models. THIS CHANGE WILL BREAK 2018.1 and below. This is now a 2019 only
rule.