Merge 2019 Dev to Main Branch (#8)
* Main Change: CityEngine 2019 GLB Vehicle Model Upgrade

These models are provided by the Esri CityEngine team as part of
CityEngine. I have used these models to replace the entire vehicles
folder. Additional models will be replaced by these GLBs with time as
they simplify the file structure, is cleaner, and easier to edit the
models. THIS CHANGE WILL BREAK 2018.1 and below. This is now a 2019 only
rule.

* Thematic setting edits

* Add Curb to Median Planters

* Update Readme in advance of 2019 (#7)

* Rotation Constant Adjustment - Vehicles

* Support New Lamps

* Remove Old Signs/Benches - Update Bench Code

* Remove old Objs

* Remove extras

* Add new traffic lights, remove old objs

* Model Change, Rule Edits  for Bus Stops

* Update Bus Stop Insertion - Sidewalk

* Add Rotation and Setback to Paint Stamp