Debugging and Changes from 2019.1 Testing
During the beta program, I worked to check if the rule works well in the
new environment. In general, it does, but during the testing process I
found various elements to update.
In the future, the current syntax for the range function updated to
require parameter declaration in the annoation.
Version was updated to 2019.1.
Sharrows had a bug removed where their size was not being consistent on
split segments.
Commented out and removed a bunch of unused elements.
Added the sidewalk bus stop setback parameter back to add movement
functionality (bug).
Freight default mix was pulling from the wrong constant. Reformated all
default mixes.
Used percent annotation in percentage based rules.
Updated scenes so there is a 2018.1 and a 2019 Scene. Removed very old
demo scene.
Updated the parklet functionality to be similar to the recent of the
percent based allocations/work with the percent annotation.
Wow this update helped reveal a lot of small things that bothered me.