Change Default Parking Length
As curbside management comes into focus and I reflect on default
assumptions, it seems that maybe the default parking length should go to
22 ft. The traffic engineer rule of thumb is one space every 25 ft give
or take, and for general applications a more conservative length for the
streets makes more sense. It might even be higher, but this achieves
immediate goals and works better with larger vehicles (pick up trucks).