Curbside Management Context / Distance Annotations
Add distance annotations to almost everything, and label curbside management features so that they can be called with context queries. For example, now at high LODs, the rule will check planting areas on sidewalks or benches on sidewalks to see if they would be too close to a drop-off zone. This should not impact performance on lower LODs.