Add Traffic Rotation
Traffic Rotation parameter added for rotating traffic signals There is a need for the handles to be revealed for translation and rotation, but one step at a time. While the way signals are by default is not realistic (opposing right side should have signal facing an opposing approach). The problem is still one of cross-segment information exchange which CGA does not allow. It might not even be possible to reliably understanding the number of links at the junction/crossing.  This is a minimal low effort change.