Freeway Rule Edits
This enables high LOD freeways. While freeways have never been a priority for the "Complete Street" rule, their integration serves a number of use cases. #28 brought attention to various old code in freeways that I am not clear if it serves a purpose anymore. There was an attempt to enable vehicles on freeways, but...this had odd results. For now, I am commenting that code snippet out...Manual placement of vehicles while not ideal is still what is required.