Address UV Flipping From Buffers
Stencils reacted strangle to flipping UVs as mentioned by #34 . This issue applied to carshare & bus stamps. It also pointed to a minor but similar issue with buffer protection & bike lane symbols (they flip when adjusted, but the arrow still points in the right direction). This seems to fix #34. For now, I am separating the UV flips from the rules calling them. This might be good for a longer term refactor.