Asset Info CSV + Rail Rule Changes (#47)
* Draft Edits - CSV Reference

Use CSVs to reference path locations per the request of CL at Esri. This seems to work on a PC test in CityEngine, but the fraction of Vans might be more explicit or half . Thank for suggestions!

* Delete GLBS (Vehicles)

New vehicles in Esri.lib are superior and are supported by the asset CSV. Delete old assets.

* Multiple Changes

Use complete folder for assets if the Asset_Metadata.csv is present. This will help force a transition.
Fix projection of rail transit pavement. Simplify rule.
TODO to fix microtransit inclusion in cars used.