Rework Parallel Parking Aligments
Based on some suggestions of Esri, I am trying to use alignScopeToGeometryBBOX for parking rather than to geometry. This could be in theory applied in other areas, but the potential impacts are unclear. This seems to provide better performance for parking spaces specifically, and the pattern used here would likely work elsewhere. Angle issues still exist at curbside drop off or longer polygons, but this seems to work better. Might for others as well that are clearly rectangular shapes.