Parking Patch (#49)
* Draft Edits - CSV Reference

Use CSVs to reference path locations per the request of CL at Esri. This seems to work on a PC test in CityEngine, but the fraction of Vans might be more explicit or half . Thank for suggestions!

* Delete GLBS (Vehicles)

New vehicles in Esri.lib are superior and are supported by the asset CSV. Delete old assets.

* Multiple Changes

Use complete folder for assets if the Asset_Metadata.csv is present. This will help force a transition.
Fix projection of rail transit pavement. Simplify rule.
TODO to fix microtransit inclusion in cars used.

* Update Asset_Metadata.csv

Reclassify transit connect vehicle as microtransit.

* Remove Duplicated Assets - Depend on Esri Lib for Street Furniture

These assets are in Esri Lib and better represented through the Asset CSV. There is some clunky behavior when selecting sidewalks and streets, but the enums do work as intended.

* Rework Parallel Parking Aligments

Based on some suggestions of Esri, I am trying to use alignScopeToGeometryBBOX for parking rather than to geometry. This could be in theory applied in other areas, but the potential impacts are unclear. This seems to provide better performance for parking spaces specifically, and the pattern used here would likely work elsewhere. Angle issues still exist at curbside drop off or longer polygons, but this seems to work better. Might for others as well that are clearly rectangular shapes.

* Rework Parallel Parking Alignments

Based on some suggestions of Esri, I am trying to use alignScopeToGeometryBBOX for parking rather than to geometry. This could be in theory applied in other areas, but the potential impacts are unclear. This seems to provide better performance for parking spaces specifically, and the pattern used here would likely work elsewhere. Angle issues still exist at curbside drop off or longer polygons, but this seems to work better. Might for others as well that are clearly rectangular shapes.

* Fix Parking

Parking is oriented in the wrong direction with align to bbox on straight streets. This addresses that but might keep some of the benefits of bbox alignment.
Minor edits to comments to remove special characters.